[fix]remove useless code

+++ yuandian-server/yuandian-netty/src/main/java/com/yuandian/server/logic/model/entity/FriendPo.java
@@ -3,15 +3,15 @@ package com.yuandian.server.logic.model.entity;
 import java.util.Date;
 
 public class FriendPo extends FriendPoKey {
-    private Integer groupId;
+    private Long groupId;
 
     private Date cTime;
 
-    public Integer getGroupId() {
+    public Long getGroupId() {
         return groupId;
     }
 
-    public void setGroupId(Integer groupId) {
+    public void setGroupId(Long groupId) {
         this.groupId = groupId;
     }
